requirements.txt
python-dotenv==1.0.1
streamlit==1.34.0
langchain==0.1.17
langchain_openai>=0.1.6
langchain_pinecone>=0.1.0
ibm-generative-ai==2.3.0
ibm-generative-ai[langchain]
momento==1.9.2